qkc/common: move Uint32 RLP codec from qkc/types (#46)
Uint32 is the fixed-width (4-byte, 0x84-prefixed) uint32 RLP form used by
pyquarkchain for full-shard-key fields. It lives in qkc/types today, but it
is a general wire-format primitive rather than a transaction/receipt type,
and consumers outside qkc/types need it: core/types needs it for the
QuarkChain 6-element account format, whose full_shard_key field is
BigEndianInt(4) in pyquarkchain's _Account.

core/types cannot import qkc/types — qkc/types already imports core/types
(bloom9.go, receipt.go), so that direction is an import cycle. qkc/common
has no such constraint and qkc/types already depends on it, so moving the
codec down is the cycle-free placement.

Pure move plus the package rename; the implementation and its golden
vectors are unchanged. qkc/types call sites now use qkcCommon.Uint32.
